+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
+### Added
+- Stats command now shows the current goroutine count and the current go runtime version
+
 ## [0.2.3] - 2019-04-13
 ### Added
 - The format for incoming lines from other games has access to a new function, `mapColours`, it allows you to map the raw